Change Grid component to TypeScript
